Title: 소괄호 매칭 수 최대화하기, Time: 1057ms, Memory: 47MB, Status: Wrong Answer - Codetree

+++ 250715/소괄호 매칭 수 최대화하기/maximize-the-number-of-parenthesis-matches.java
@@ -16,8 +16,8 @@ public class Main {
             sb.append(s);
         }
         String str = sb.toString();
-        int cnt = 0;
-        int back = 0;
+        long cnt = 0;
+        long back = 0;
         for (int i = str.length() - 1; i >= 0; i--){
             if (str.charAt(i) == ')')
                 back++;
